internal/e2e: add a "cue mod upload" end-to-end test with gcloud
Now that we have the first experimental version of "cue mod upload",
and we support authenticating with cloud provider registries
via docker credential helpers such as gcloud,
we can write an end-to-end test with Google Cloud Artifact Registry.

This test does the same as our existing GitHub App test at a high level,
publishing one module version and then using it as a dependency.
However, instead of creating a GitHub repo and pushing a git tag,
we use "cue mod upload" with gcloud as a credential helper,
making sure that gcloud shares the same configuration as the host.

The test script is named after "gcloud" since all the auth bits
are particular to Google Cloud right now.
We can later add another script, or make this script more generic,
if we want to test against other common OCI/Docker registry providers.

Note that the test is skipped for now for the sake of our CI,
since we don't have a service account set up with a key on it yet.
The next commit will do that bit of setup in our CI script.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: I2e45865bf86788c42ba3ef2d7422fd21b9af221a